(fix) Changed the logic to configure the list of accepted spot connectors in the base config map class for strategies
