+++ bin/hummingbot_quickstart.py
@@ -25,6 +25,7 @@ from hummingbot.client.config.config_helpers import (
     read_system_configs_from_yml,
 )
 from hummingbot.client.config.security import Security
+from hummingbot.client.config.strategy_config_data_types import BaseStrategyConfigMap
 from hummingbot.client.hummingbot_application import HummingbotApplication
 from hummingbot.client.settings import STRATEGIES_CONF_DIR_PATH, AllConnectorSettings
 from hummingbot.client.ui import login_prompt
