+++ hummingbot/client/settings.py
@@ -412,7 +412,7 @@ class AllConnectorSettings:
     @classmethod
     def get_exchange_names(cls) -> Set[str]:
         return {
-            cs.name for cs in cls.all_connector_settings.values() if cs.type is ConnectorType.Exchange
+            cs.name for cs in cls.get_connector_settings().values() if cs.type is ConnectorType.Exchange
         }.union(set(PAPER_TRADE_EXCHANGES))
 
     @classmethod
